controller:lib: Fixed type for AEM and ACMP command status. Added more detailed descriptor for notification and logging. Updated notification callback to include command status.

+++ controller/lib/src/util_imp.cpp
@@ -338,7 +338,7 @@ namespace avdecc_lib
         return (uint16_t)AEM_DESC_ERROR;
     }
 
-    const char * STDCALL util_imp::aem_cmd_status_value_to_name(uint16_t aem_cmd_status_value)
+    const char * STDCALL util_imp::aem_cmd_status_value_to_name(uint32_t aem_cmd_status_value)
     {
         if(aem_cmd_status_value < TOTAL_NUM_OF_AEM_CMDS_STATUS)
         {
@@ -384,7 +384,7 @@ namespace avdecc_lib
         return (uint16_t)AEM_CMD_ERROR;
     }
 
-    const char * STDCALL util_imp::acmp_cmd_status_value_to_name(uint16_t acmp_cmd_status_value)
+    const char * STDCALL util_imp::acmp_cmd_status_value_to_name(uint32_t acmp_cmd_status_value)
     {
         if(acmp_cmd_status_value < TOTAL_NUM_OF_ACMP_CMDS_STATUS)
         {
